Separate span lifecycle from exec_id correlation in the tracing tests
`test_tracing_adapter.py` had grown to 623 lines and was carrying three
unrelated concerns. Move the span-lifecycle coverage — the ancillary
phases, the late `teardown_error`, and both eviction tests — into
`test_tracing_span_lifecycle.py`, along with the `_cat_overrides` helper
they are the only users of.

Two modules could not hold the result: the pair totalled 1003 lines
against a 400-line ceiling each, so the arithmetic never worked. The
correlation tests are the natural third concern — they are about *which*
span an event belongs to, not what happens to a span between `start` and
`exit` — so they move to `test_tracing_exec_id_correlation.py` together
with `test_events_without_exec_id_are_ignored`, which was answering the
same question from the adapter module.

Every test keeps its assertions and its shared `exec_id` setup; the three
modules collect exactly the 27 tests the two did before. Command
execution, span attributes, the factory, the in-memory tracer, and the
concurrency coverage stay put.
